requirements.txt
black>=24.0.0
ruff>=0.4.0
anthropic>=0.25.0
openai>=1.0.0
tiktoken>=0.6.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-cov>=4.0.0
google-generativeai>=0.4.0
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
